Split enface visibility helpers and shrink projection manager.
Move layer visibility, alpha, and color logic to enfaceProjectionVisibility.ts with enumerateVisibleEnfaceLayers.
